LayoutTests/platform/iphone-simulator/accessibility/link-with-only-image-expected.txt

 
This test makes sure that a link with only an image will report the correct label.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.iphoneIsElement is true
PASS obj.iphoneLabel is 'AXLabel: Hello World'
PASS successfullyParsed is true

TEST COMPLETE
